Fixed error in parsing energy resolution to STIR
Energy resolution in the scanner parameter file must be defined as a percentage (%), as gathered by SimSET: https://depts.washington.edu/simset/html/user_guide/detector_modelling.html

On the other hand, STIR requires the energy resolution to be expressed as a fraction of the reference energy (which is set to 511). This means that instead of a perfectage needs to be expressed as a fraction of 1.

So the percentage needs to be divided by 100 to be fraction of 1.