Updated simset_sim
Now, second sample simulation (for estimating n_photons) will only be performed if importance sampling is activated. This solves a problem when the user does not want to use importance sampling.

Additionally, I entered a condition that will only trigger this calculation if the user has not stated photons manually.

So, correct way of launching simulations with importance sampling and automatic calculations should be:

s_photons = 1543214 (int num)
photons = 0

This is well matched with photons = 0  and s_photons=0 triggers autocalculation (from simset) and no importance sampling